requirements.txt
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
cycler==0.11.0
fonttools==4.28.5
h11==0.12.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
outcome==1.1.0
packaging==21.3
Pillow==8.4.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytesseract==0.3.8
python-dateutil==2.8.2
python-decouple==3.5
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
